Initial layout
Added the basic method of importing and showing cluster data. I tried to
keep the meat of the algorithm (AgentCluster) as STL-based as possible,
while keeping the visualization part (ClusterCanvas) more Qt-based. This
separation will hopefully make individual pieces of the application
easier to move around and use for other projects.