Basic movement functionality, new datasets.
Managed to iron out some of the blaring bugs, and it looks like the
agents are at least moving towards the clusters. However, I cheated and
just used a very simple set of linear functions to calculate the
happiness value, while I eventually want to use logistic-style functions
for that.

I do notice that both the effective and personal space range of agents
will shrink a lot when it seems to "settle" in on a point. I have a
feeling this has to do with two agents occupying the same position. I
will probably need to add some functionality to prevent having two
agents occupy the same position.

I also added some new data sets for testing.