Added support for general FASO use.
Tries to find the global minimum of a Styblinski–Tang function. Draws on
a contour plot image for visual inspection.Right now, it looks like it's
trying to find the global max...need to fix this. The drawing framework
is down pretty well, it's just a matter of tweaking the FASO class to
better suit the gradient based function. To use general FASO mode, just
don't pass "-c" as a command line parameter.