Fix FASO mode, added another test function.
Just needed to fix the objective function, as well as the gradient
function. Added support for multiple test functions, and currently
includes the Styblinski and Ackley function. It's simple enough to add
more later.

Note: One issue I noticed is that positioning on the Styblinski plot is
not quite right; I exported the final x/y positions of the agents to a
csv file, and they're correctly covering the minima. Just looks like I'm
not translating them to the canvas correctly.